コントローラサービス用プロセスクラス群作成

+++ src/Lib/UUID/UUID.php
@@ -103,7 +103,10 @@ class UUID
      */
     private static int $unixts_ms = 0;
 
-    /** @internal */
+    /**
+     * @internal
+     * @return int[] [$unixts, $subsec]
+     */
     private static function getUnixTimeSubsec(): array
     {
         $timestamp = microtime(false);
@@ -342,7 +345,7 @@ class UUID
             $retval = substr_replace(str_pad($retval, 8, '0', STR_PAD_LEFT), '.', -7, 0);
         } elseif ($version === 8) {
             $unixts = hexdec(substr($timehex, 0, 13));
-            $subsec = self::decodeSubsec((hexdec(substr($timehex, 13)) << 2) + (hexdec(substr($uuid, 16, 1)) & 0x03));
+            $subsec = self::decodeSubsec((int) (hexdec(substr($timehex, 13)) << 2) + (hexdec(substr($uuid, 16, 1)) & 0x03));
             $retval = strval($unixts * self::V8_SUBSEC_RANGE + $subsec);
             $retval = substr_replace(str_pad($retval, 8, '0', STR_PAD_LEFT), '.', -7, 0);
         }
